# docs-v2/30-api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Teranga API",
    "description": "African Event Management Platform — REST API for events, registrations, badges, and users",
    "version": "0.1.0",
    "contact": {
      "name": "Teranga Team",
      "email": "dev@teranga.events"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT",
        "description": "Firebase ID token obtained via Firebase Auth SDK"
      },
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "terk_<env>_<40chars>_<4chksum>",
        "description": "Organization API key (T2.3). Issued via /admin/organizations/[id]?tab=api-keys."
      }
    },
...
  },
  "paths": {
...
    "/v1/me/whatsapp/opt-in": {
      "post": {
        "summary": "Grant WhatsApp opt-in for the calling user (org-scoped)",
        "tags": [
          "WhatsApp"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Revoke a previously-granted WhatsApp opt-in",
        "tags": [
          "WhatsApp"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "summary": "Read the current WhatsApp opt-in record (or null)",
        "tags": [
          "WhatsApp"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
}